test: consolidate observability tests into test_observability_basic.py
- Merged test_observability_simple.py and test_observability_working.py
- Created unified test_observability_basic.py with 14 tests (all passing)
- Organized into 2 test classes:
  - TestObservabilityBasic: basic tests without setup/teardown
  - TestObservabilityWithSetup: tests requiring temp directory
- Fixed metadata expectations and mypy type checking
- Removed temporary editor file

Reduces test duplication while maintaining comprehensive coverage.

Also includes previously committed changes:
- Updated pre-commit hooks to latest versions
- Cleaned up unused import in test_pipeline_integration.py
