+++ .pre-commit-config.yaml
@@ -1,14 +1,14 @@
 repos:
   # General file cleanup
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v5.0.0
+    rev: v6.0.0
     hooks:
       - id: end-of-file-fixer
       - id: trailing-whitespace
 
   # Black - Python code formatter
   - repo: https://github.com/psf/black
-    rev: 25.1.0
+    rev: 25.9.0
     hooks:
       - id: black
         language_version: python3
@@ -17,7 +17,7 @@ repos:
 
   # isort - Python import sorter
   - repo: https://github.com/pycqa/isort
-    rev: 6.0.1
+    rev: 7.0.0
     hooks:
       - id: isort
         args: [--profile=black, --line-length=100]
@@ -33,7 +33,7 @@ repos:
 
   # MyPy - Python type checker
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.17.1
+    rev: v1.18.2
     hooks:
       - id: mypy
         additional_dependencies: [types-requests, types-PyYAML, types-toml]
