--- tests/test_observability_working.py
@@ -1,299 +1,0 @@
-"""
-Working tests for observability integration that focus on actual functionality.
-"""
-
-import os
-import shutil
-import tempfile
-from unittest.mock import MagicMock, patch
-
-from sentries.chat import chat
-
-
-class TestObservabilityWorking:
-    """Tests that focus on working observability functionality."""
-
-    def setup_method(self) -> None:
-        """Set up test environment."""
-        self.test_dir = tempfile.mkdtemp()
-        self.original_cwd = os.getcwd()
-        os.chdir(self.test_dir)
-
-    def teardown_method(self) -> None:
-        """Clean up test environment."""
-        os.chdir(self.original_cwd)
-        shutil.rmtree(self.test_dir)
-
-    def test_observability_integration_basic(self) -> None:
-        """Test basic observability integration works."""
-        with patch.dict(os.environ, {"SENTRIES_SIMULATION_MODE": "true"}):
-            mock_log = MagicMock()
-            mock_analyze = MagicMock(return_value={"pii_spans": []})
-
-            with patch("packages.metrics_core.observability.log_llm_interaction", mock_log):
-                with patch(
-                    "packages.metrics_core.observability.analyze_text_for_pii", mock_analyze
-                ):
-                    messages = [{"role": "user", "content": "Hello"}]
-                    response = chat("test-model", messages)
-
-                    # Verify observability was called
-                    assert mock_log.called
-                    assert mock_analyze.called
-                    assert isinstance(response, str)
-
-                    # Check basic metadata
-                    call_args = mock_log.call_args
-                    metadata = call_args[1]["metadata"]
-                    assert metadata["mode"] == "simulation"
-                    assert metadata["model"] == "test-model"
-
-                    print("✅ Basic observability integration working")
-
-    def test_observability_class_integration(self) -> None:
-        """Test observability class integration directly."""
-        try:
-            from packages.metrics_core.observability import TestSentryObservability
-
-            obs = TestSentryObservability()
-
-            # Test logging
-            obs.log_llm_interaction(
-                prompt="Test prompt",
-                response="Test response",
-                metadata={"mode": "simulation", "model": "test-model", "temperature": 0.5},
-            )
-
-            # Check that event was stored
-            assert len(obs.current_events) == 1
-            event = obs.current_events[0]
-
-            # Check enhanced metadata
-            assert event.metadata["mode"] == "simulation"
-            assert event.metadata["is_simulation"] is True
-            assert event.metadata["is_api"] is False
-            assert event.metadata["is_local"] is False
-            assert event.metadata["model"] == "test-model"
-            assert event.metadata["temperature"] == 0.5
-
-            print("✅ Observability class integration working")
-
-        except ImportError as e:
-            print(f"⚠️  Observability class import issue: {e}")
-
-    def test_pii_analysis_integration(self) -> None:
-        """Test PII analysis integration."""
-        try:
-            from packages.metrics_core.observability import TestSentryObservability
-
-            obs = TestSentryObservability()
-
-            # Test PII analysis
-            test_text = "Contact me at john.doe@example.com"
-            result = obs.analyze_pii_in_text(test_text)
-
-            assert "pii_spans" in result
-            assert "pii_stats" in result
-            assert isinstance(result["pii_spans"], list)
-
-            print("✅ PII analysis integration working")
-
-        except ImportError as e:
-            print(f"⚠️  PII analysis import issue: {e}")
-        except Exception as e:
-            print(f"⚠️  PII analysis runtime issue: {e}")
-
-    def test_mode_detection_in_chat(self) -> None:
-        """Test that mode detection works correctly in chat function."""
-        test_scenarios = [
-            ("simulation", {"SENTRIES_SIMULATION_MODE": "true"}),
-            ("api", {"GROQ_API_KEY": "test-key"}),
-            ("local", {}),
-        ]
-
-        for expected_mode, env_vars in test_scenarios:
-            with patch.dict(os.environ, env_vars, clear=True):
-                mock_log = MagicMock()
-                mock_analyze = MagicMock(return_value={"pii_spans": []})
-
-                # Mock appropriate chat function
-                if expected_mode == "api":
-                    mock_target = "sentries.chat.chat_with_groq"
-                elif expected_mode == "local":
-                    mock_target = "sentries.chat.chat_with_ollama"
-                else:
-                    mock_target = None
-
-                mock_context = (
-                    patch(mock_target, return_value="Test response") if mock_target else MagicMock()
-                )
-
-                with mock_context:
-                    with patch("packages.metrics_core.observability.log_llm_interaction", mock_log):
-                        with patch(
-                            "packages.metrics_core.observability.analyze_text_for_pii", mock_analyze
-                        ):
-                            messages = [{"role": "user", "content": "Hello"}]
-                            chat("test-model", messages)
-
-                            # Verify mode was detected correctly
-                            assert mock_log.called
-                            metadata = mock_log.call_args[1]["metadata"]
-                            assert metadata["mode"] == expected_mode
-
-                            print(f"✅ Mode detection working for {expected_mode}")
-
-    def test_observability_error_handling(self) -> None:
-        """Test that observability errors don't break chat."""
-        with patch.dict(os.environ, {"SENTRIES_SIMULATION_MODE": "true"}):
-            # Test with observability import error
-            def mock_import_error(*args, **kwargs):
-                raise ImportError("Observability not available")
-
-            with patch(
-                "packages.metrics_core.observability.log_llm_interaction",
-                side_effect=mock_import_error,
-            ):
-                messages = [{"role": "user", "content": "Hello"}]
-                response = chat("test-model", messages)
-
-                # Should still work
-                assert isinstance(response, str)
-                assert len(response) > 0
-
-                print("✅ Error handling working")
-
-    def test_observability_performance(self) -> None:
-        """Test observability performance impact."""
-        import time
-
-        with patch.dict(os.environ, {"SENTRIES_SIMULATION_MODE": "true"}):
-            messages = [{"role": "user", "content": "Performance test"}]
-
-            # Test without observability
-            with patch(
-                "packages.metrics_core.observability.log_llm_interaction", side_effect=ImportError
-            ):
-                start_time = time.time()
-                for _ in range(5):
-                    chat("test-model", messages)
-                without_obs_time = time.time() - start_time
-
-            # Test with observability (mocked)
-            mock_log = MagicMock()
-            mock_analyze = MagicMock(return_value={"pii_spans": []})
-
-            with patch("packages.metrics_core.observability.log_llm_interaction", mock_log):
-                with patch(
-                    "packages.metrics_core.observability.analyze_text_for_pii", mock_analyze
-                ):
-                    start_time = time.time()
-                    for _ in range(5):
-                        chat("test-model", messages)
-                    with_obs_time = time.time() - start_time
-
-            # Calculate overhead
-            overhead = with_obs_time - without_obs_time
-            overhead_per_call = overhead / 5
-
-            print(f"✅ Performance overhead: {overhead_per_call:.4f}s per call")
-            assert overhead_per_call < 0.1  # Should be minimal
-
-    def test_full_integration_with_real_observability(self) -> None:
-        """Test full integration with real observability components."""
-        try:
-            from packages.metrics_core.observability import get_observability
-
-            with patch.dict(os.environ, {"SENTRIES_SIMULATION_MODE": "true"}):
-                # Get the real observability instance
-                obs = get_observability()
-                initial_events = len(obs.current_events)
-
-                # Make a chat call
-                messages = [{"role": "user", "content": "Test real observability"}]
-                chat("test-model", messages, temperature=0.7)
-
-                # Check that event was logged
-                final_events = len(obs.current_events)
-                assert final_events > initial_events
-
-                # Check the logged event
-                latest_event = obs.current_events[-1]
-                assert latest_event.event_type == "llm_interaction"
-                assert "Test real observability" in latest_event.message
-                assert latest_event.metadata["mode"] == "simulation"
-                assert latest_event.metadata["is_simulation"] is True
-                assert latest_event.metadata["model"] == "test-model"
-                assert latest_event.metadata["temperature"] == 0.7
-
-                print("✅ Full real observability integration working")
-
-        except ImportError as e:
-            print(f"⚠️  Real observability import issue: {e}")
-        except Exception as e:
-            print(f"⚠️  Real observability runtime issue: {e}")
-
-    def test_backward_compatibility(self) -> None:
-        """Test backward compatibility with chat_with_observability."""
-        from sentries.testsentry import chat_with_observability
-
-        with patch.dict(os.environ, {"SENTRIES_SIMULATION_MODE": "true"}):
-            mock_log = MagicMock()
-            mock_analyze = MagicMock(return_value={"pii_spans": []})
-
-            with patch("packages.metrics_core.observability.log_llm_interaction", mock_log):
-                with patch(
-                    "packages.metrics_core.observability.analyze_text_for_pii", mock_analyze
-                ):
-                    messages = [{"role": "user", "content": "Backward compatibility test"}]
-                    response = chat_with_observability("test-model", messages)
-
-                    # Should work and call observability
-                    assert isinstance(response, str)
-                    assert mock_log.called
-
-                    print("✅ Backward compatibility working")
-
-    def test_concurrent_observability(self) -> None:
-        """Test observability with concurrent requests."""
-        import threading
-        import time
-
-        with patch.dict(os.environ, {"SENTRIES_SIMULATION_MODE": "true"}):
-            mock_log = MagicMock()
-            mock_analyze = MagicMock(return_value={"pii_spans": []})
-            results = []
-            errors = []
-
-            def make_request(thread_id: int) -> None:
-                try:
-                    with patch("packages.metrics_core.observability.log_llm_interaction", mock_log):
-                        with patch(
-                            "packages.metrics_core.observability.analyze_text_for_pii", mock_analyze
-                        ):
-                            messages = [{"role": "user", "content": f"Thread {thread_id}"}]
-                            response = chat(f"model-{thread_id}", messages)
-                            results.append((thread_id, response))
-                except Exception as e:
-                    errors.append((thread_id, str(e)))
-
-            # Create and run threads
-            threads = []
-            for i in range(3):
-                thread = threading.Thread(target=make_request, args=(i,))
-                threads.append(thread)
-
-            start_time = time.time()
-            for thread in threads:
-                thread.start()
-
-            for thread in threads:
-                thread.join()
-            end_time = time.time()
-
-            # Check results
-            assert len(errors) == 0, f"Errors: {errors}"
-            assert len(results) == 3
-
-            total_time = end_time - start_time
-            print(f"✅ Concurrent observability working: {total_time:.2f}s for 3 threads")
