+++ tests/test_pipeline_integration.py
@@ -14,8 +14,6 @@ import pytest
 
 from sentries.chat import chat, has_api_key, is_simulation_mode
 
-# from sentries.testsentry import main as testsentry_main  # Not used in current tests
-
 
 class TestSimulationPipeline:
     """Test complete TestSentry pipeline in simulation mode."""
